+++ packages/features/form-builder/index.ts
@@ -1,5 +1,6 @@
 // TODO: FormBuilder makes more sense in @calcom/ui but it has an additional thing that other components don't have
 // It has zod schema associated with it and I currently can't import zod in there.
 // Move it later there maybe? @sean
-export { FormBuilder } from "./FormBuilder";
+export { FormBuilder, FieldEditDialogForm, RhfFormField } from "./FormBuilder";
 export { FormBuilderField } from "./FormBuilderField";
+export { AddQuestionsForm } from "./AddQuestionsForm";
